Add an ui for role editing (#20)
* Begin role ui implementation

This is by no means final, just the start of an implementation for the roles ui

* Change how role panel is opened

* Add method to permissions to get list of roles

This was pretty fun to program.

Created a new simple class, "Role", as a simple abstraction to make it easier for external callers to interact with the role. (This will be a very helpful class for the UI)

* Fix errors getting all permissions

1st, the method should have been returning the values, not the dict we were using to build the list of values.
2nd, the brackets operator actually errors if we try retrieving a value that doesn't exist. Very curious, considering assignment still works. The .get() method is safer, it returns `None` by default for undefined properties.
3rd, it should have been "can_modify_roles", not "can_add_role" in the constructor. Oops.

* Add repo to roles constructor | get actual list of roles for repo

I had to move the manage roles button from the header and add it to the repo list. I didn't realize the roles list would need to be repo-sensitive at first, but it needs to be.

Also, add logic to get the real list of roles from the repo. It's great - everything's working now!

* Fix rebase conflicts

* Add the roleeditpanel

Not fully functional yet, but it looks almost fully functional.

* Actually use the permissions list from the list of permissions

* edit a copy of the permissions, not the actual permissions itself

* Undo all changes to permissions file

* Removed unused roles panel

* Change roles editor to not use deleted get_roles method

This means that the roles editor is now completely relying on the permissions file

* Check the permissions on the actual user

* Clean the lint trap